Files for a context-free grammar parser
Final project of Data Structures & Algorithms. I wrote the all the code in the FileReader file, and about half the code in the RandomPhraseGenerator file along with a partner (paired programming). This code is meant to take in a specifically structured grammar file and output a random phrase that is grammatically correct according to the rules set out in a .g file. "abc.g" file is provided to show what is in a very basic .g file to be parsed.